Mp 25 (#36)
* nochange

* boilerplate for transaction backend

* finishing backend order

* add toast in cart

* add next page if some is empty

* claer

* report transaction list + filter by date

* report chart

* before merge

* after merge

* fix some bug

* fix order change

* add something on modal order

* some

* refetching after update - prroduct

* fix category update

* merge

* after merge

+++ projects/server/src/controllers/productController.js
@@ -28,7 +28,7 @@ async function addProduct(req, res) {
         const result = await productService.addProduct(req);
         res.status(result.status).json(messages.response(result));
     } catch (error) {
-        await fs.unlink(req.file.path);
+        if (req.file) await fs.unlink(req.file.path);
         res.status(500).json({ message: error.message });
     }
 }
